Add support for Doctrine mapped superclasses
